requirements.txt
apache-airflow==2.7.3
astronomer-providers==1.11.0
pandas==2.1.3
requests==2.31.0
psycopg2-binary==2.9.9
dvc==3.40.0
dvc[s3]==3.40.0
python-dotenv==1.0.0

